+++ pwsh/AzGovVizParallel.ps1
@@ -362,7 +362,7 @@ Param
     $AzAPICallVersion = '1.1.54',
 
     [string]
-    $ProductVersion = 'v6_major_20221121_1',
+    $ProductVersion = 'v6_major_20221128_1',
 
     [string]
     $GithubRepository = 'aka.ms/AzGovViz',
@@ -448,7 +448,7 @@ Param
 
     [Alias('ExludedResourceTypesDiagnosticsCapable')]
     [array]
-    $ExcludedResourceTypesDiagnosticsCapable = @('microsoft.web/certificates'),
+    $ExcludedResourceTypesDiagnosticsCapable = @('microsoft.web/certificates', 'microsoft.chaos/chaosexperiments'),
 
     [switch]
     $DoNotIncludeResourceGroupsOnPolicy,
@@ -7588,7 +7588,22 @@ function processPrivateEndpoints {
             $SubnetVNetResourceGroup = $hlper.ResourceGroup
         }
 
-        $resourceSubscriptionId = ($pe.properties.privateLinkServiceConnections.properties.privateLinkServiceId -split '/')[2]
+        if ($pe.properties.privateLinkServiceConnections.Count -gt 0) {
+            $resourceId = $pe.properties.privateLinkServiceConnections.properties.privateLinkServiceId
+            $targetSubresource = $pe.properties.privateLinkServiceConnections.properties.groupIds -join ', '
+            $resourceSplit = $pe.properties.privateLinkServiceConnections.properties.privateLinkServiceId -split '/'
+            $peConnectionType = 'direct'
+            $peConnectionState = $pe.properties.privateLinkServiceConnections.properties.privateLinkServiceConnectionState.status
+        }
+        if ($pe.properties.manualPrivateLinkServiceConnections.Count -gt 0) {
+            $resourceId = $pe.properties.manualPrivateLinkServiceConnections.properties.privateLinkServiceId
+            $targetSubresource = $pe.properties.manualPrivateLinkServiceConnections.properties.groupIds -join ', '
+            $resourceSplit = $pe.properties.manualPrivateLinkServiceConnections.properties.privateLinkServiceId -split '/'
+            $peConnectionType = 'manual'
+            $peConnectionState = $pe.properties.manualPrivateLinkServiceConnections.properties.privateLinkServiceConnectionState.status
+        }
+
+        $resourceSubscriptionId = $resourceSplit[2]
         $resourceSubscriptionName = 'n/a'
         $resourceMGPath = 'n/a'
         if ($htSubscriptionsMgPath.($resourceSubscriptionId)) {
@@ -7604,7 +7619,7 @@ function processPrivateEndpoints {
             $crossSubscriptionPE = $true
         }
 
-        $resourceSplit = $pe.properties.privateLinkServiceConnections.properties.privateLinkServiceId -split '/'
+
 
         $null = $script:arrayPrivateEndpointsEnriched.Add([PSCustomObject]@{
                 PEName                   = $pe.name
@@ -7614,12 +7629,14 @@ function processPrivateEndpoints {
                 PESubscriptionName       = $subscriptionName
                 PESubscription           = ($pe.id -split '/')[2]
                 PEMGPath                 = $MGPath
+                PEConnectionType         = $peConnectionType
+                PEConnectionState        = $peConnectionState
                 CrossSubscriptionPE      = $crossSubscriptionPE
 
                 Resource                 = $resourceSplit[8]
                 ResourceType             = "$($resourceSplit[6])/$($resourceSplit[7])"
-                ResourceId               = $pe.properties.privateLinkServiceConnections.properties.privateLinkServiceId
-                TargetSubresource        = $pe.properties.privateLinkServiceConnections.properties.groupIds -join ', '
+                ResourceId               = $resourceId
+                TargetSubresource        = $targetSubresource -join ', '
                 NICName                  = $pe.properties.customNetworkInterfaceName
                 FQDN                     = $pe.properties.customDnsConfigs.fqdn -join ', '
                 ipAddresses              = $pe.properties.customDnsConfigs.ipAddresses -join ', '
@@ -21542,6 +21559,8 @@ btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { 
 <th>PE Subscription Name</th>
 <th>PE Subscription</th>
 <th>PE MGPath</th>
+<th>PE Type</th>
+<th>PE State</th>
 <th>Cross Subscription PE</th>
 
 <th class="uamiresaltbgc">Resource</th>
@@ -21581,7 +21600,9 @@ btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { 
                         <td>$($result.PESubscriptionName)</td>
                         <td>$($result.PESubscription)</td>
                         <td style="min-width: 150px" class="breakwordall">$($result.PEMGPath)</td>
-                        <td>$($result.crossSubscriptionPE)</td>
+                        <td>$($result.PEConnectionType)</td>
+                        <td>$($result.PEConnectionState)</td>
+                        <td>$($result.CrossSubscriptionPE)</td>
 
                         <td>$($result.Resource)</td>
                         <td>$($result.ResourceType)</td>
@@ -21647,10 +21668,14 @@ btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { 
             linked_filters: true,
             col_2: 'select',
             col_7: 'select',
+            col_8: 'select',
             col_9: 'select',
             col_11: 'select',
-            col_23: 'select',
+            col_13: 'select',
+            col_25: 'select',
             col_types: [
+                'caseinsensitivestring',
+                'caseinsensitivestring',
                 'caseinsensitivestring',
                 'caseinsensitivestring',
                 'caseinsensitivestring',